+++ Rec/Muon/MuonTrackAlign/src/AlignMuonTrack.h
@@ -1,4 +1,4 @@
-// $Id: AlignMuonTrack.h,v 1.1.1.1 2007-07-23 07:14:36 asatta Exp $
+// $Id: AlignMuonTrack.h,v 1.2 2010-03-22 01:20:06 rlambert Exp $
 #ifndef ALIGNMUONTRACK_H 
 #define ALIGNMUONTRACK_H 1
 
@@ -43,12 +43,12 @@ public:
   
   float slopeX(int stationFirst,int stationSecond,float zFirst, float zSecond){
     return 
-      (m_points[stationFirst][0].x()-m_points[stationSecond][0].x())/(zFirst-zSecond);
+      float((m_points[stationFirst][0].x()-m_points[stationSecond][0].x())/(zFirst-zSecond));
   };
   float slopeY(int stationFirst,int stationSecond, 
                float zFirst, float zSecond){
     return 
-      (m_points[stationFirst][0].y()-m_points[stationSecond][0].y())/(zFirst-zSecond);
+      float((m_points[stationFirst][0].y()-m_points[stationSecond][0].y())/(zFirst-zSecond));
   };
   
   float slopeX(int stationFirst,int stationSecond){
@@ -57,7 +57,7 @@ public:
   };
   float slopeY(int stationFirst,int stationSecond){
     return 
-      (m_points[stationFirst][0].y()-m_points[stationSecond][0].y())/(m_points[stationFirst][0].z()-m_points[stationSecond][0].z());
+      float((m_points[stationFirst][0].y()-m_points[stationSecond][0].y())/(m_points[stationFirst][0].z()-m_points[stationSecond][0].z()));
   };
